% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-internal.R
\name{absoluteGrob}
\alias{absoluteGrob}
\title{Absolute grob}
\usage{
absoluteGrob(
  grob,
  width = NULL,
  height = NULL,
  xmin = NULL,
  ymin = NULL,
  vp = NULL
)
}
\description{
This grob has fixed dimensions and position.
}
\details{
It's still experimental
}
\keyword{internal}
