% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pvalue.R
\name{add_pvalue}
\alias{add_pvalue}
\title{Title}
\usage{
add_pvalue(
  data,
  label = NULL,
  xmin = "group1",
  xmax = "group2",
  x = NULL,
  y.position = "y.position",
  label.size = 3.2,
  colour = NULL,
  color = NULL,
  tip.length = 0.03,
  bracket.size = 0.6,
  bracket.colour = NULL,
  bracket.shorten = 0,
  bracket.nudge.y = 0,
  step.increase = 0,
  step.group.by = NULL,
  remove.bracket = FALSE,
  coord.flip = FALSE,
  position = "identity",
  ...
)
}
\arguments{
\item{data}{Description.}

\item{label}{Description.}

\item{xmin}{Description.}

\item{xmax}{Description.}

\item{x}{Description.}

\item{y.position}{Description.}

\item{label.size}{Description.}

\item{colour}{Description.}

\item{color}{Description.}

\item{tip.length}{Description.}

\item{bracket.size}{Description.}

\item{bracket.colour}{Description.}

\item{bracket.shorten}{Description.}

\item{bracket.nudge.y}{Description.}

\item{step.increase}{Description.}

\item{step.group.by}{Description.}

\item{remove.bracket}{Description.}

\item{coord.flip}{Description.}

\item{position}{Description.}

\item{...}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{
#

}
