% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_ticks.R
\name{annotation_ticks}
\alias{annotation_ticks}
\title{Add ticks as ggplot annotation}
\source{
The code is a slightly modified version of the answer to this
[Stack Overflow](https://stackoverflow.com/questions/58485334) question,
which is itself a refactored version of this
[annotation_ticks](https://github.com/hrbrmstr/ggalt/blob/master/R/annotation_ticks.r)
function.
}
\usage{
annotation_ticks(
  sides = "b",
  type = "both",
  scale = "identity",
  scaled = TRUE,
  outside = FALSE,
  tick.length = unit(4.8, "pt"),
  minor.length = unit(2.4, "pt"),
  size = 0.6,
  colour = "black",
  color = NULL,
  linetype = 1,
  lineend = "butt",
  alpha = 1,
  data = data.frame(x = NA)
)
}
\arguments{
\item{sides}{`string`. Indicates which sides of the plot should ticks
appear. Can be any of `"trbl"`, for top, right, bottom, left.}

\item{type}{`string`. Types of ticks that appear. One of
`"major"`, `"minor"`, or `"both"`. Control number of ticks
by controlling the `breaks` and `minor_breaks` arguments in the
various ggplot2 `scale_(x|y)_` functions.}

\item{scale}{`string` or `character vector`. Type of scale to be
used for each side, for example `"identity"` or `"log10"`. If
length is 1 then the scale will be used for all `sides`. If longer than
1 the length must match the number of `sides` specified.}

\item{scaled}{`logical`. Is the data already scaled? Should be
`TRUE` (default) if data is already transformed with `log10()` or
when using `scale_y_log10()`. Should be `FALSE` when using
`coord_trans(y = "log10")`.}

\item{outside}{`logical`. Should the ticks point outside of the plotting
area? If `TRUE` clipping must be turned off.}

\item{tick.length}{a \code{\link[grid]{unit}} object specifying the length
of major ticks.}

\item{minor.length}{a \code{\link[grid]{unit}} object specifying the length
of mminor ticks.}

\item{size}{`numeric`. Linewidth of ticks.}

\item{colour, color}{`string`. Colour of ticks.}

\item{linetype}{`string` or `numeric`. Linetype of tick marks.}

\item{lineend}{`string`. Lineend of ticks. One of `"square"`
(default), `"butt"`, or `"round"`.}

\item{alpha}{`numeric`. Transparency of ticks.}

\item{data}{`data.frame`. Use this argument to control the appearance of
ticks on different facets. Pass a data.frame containing the levels from the
faceting variable you want to annotate specifically.
See [here](https://stackoverflow.com/questions/20128582) for an example.}
}
\description{
This is an annotation function to add tick marks (major, minor, or both) to
a ggplot. Clipping must be turned off if the ticks are to appear outside the
plotting area, for example with: `coord_cartesian(clip = "off")`.
}
\examples{
# example goes here
}
