% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_prism_offset_minor.R
\name{draw_prism_offset_minor}
\alias{draw_prism_offset_minor}
\title{Grob for offset axes with minor ticks}
\usage{
draw_prism_offset_minor(
  break_positions,
  break_labels,
  breaks_major,
  axis_position,
  theme,
  check.overlap = FALSE,
  angle = NULL,
  n.dodge = 1
)
}
\arguments{
\item{break_positions}{position of ticks}

\item{break_labels}{labels at ticks}

\item{breaks_major}{logical vector indicating major ticks versus minor ticks}

\item{axis_position}{position of axis (top, bottom, left or right)}

\item{theme}{A complete \code{\link[ggplot2]{theme}} object}

\item{check.overlap}{silently remove overlapping labels,
(recursively) prioritizing the first, last, and middle labels.}

\item{angle}{Compared to setting the angle in
\code{\link[ggplot2]{theme}} / \code{\link[ggplot2]{element_text}},
this also uses some heuristics to automatically pick the \code{hjust} and
\code{vjust} that you probably want.}

\item{n.dodge}{The number of rows (for vertical axes) or columns (for
horizontal axes) that should be used to render the labels. This is
useful for displaying labels that would otherwise overlap.}
}
\description{
Grob for offset axes with minor ticks
}
\keyword{internal}
