% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bracket.R
\name{geom_bracket}
\alias{geom_bracket}
\title{Title}
\usage{
geom_bracket(
  mapping = NULL,
  data = NULL,
  stat = "bracket",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  label = NULL,
  type = c("text", "expression"),
  xmin = NULL,
  xmax = NULL,
  y.position = NULL,
  label.size = 3.2,
  tip.length = 0.03,
  bracket.size = 0.6,
  bracket.colour = NULL,
  bracket.shorten = 0,
  bracket.nudge.y = 0,
  step.increase = 0,
  step.group.by = NULL,
  coord.flip = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Description.}

\item{data}{Description.}

\item{stat}{Description.}

\item{position}{Description.}

\item{na.rm}{Description.}

\item{show.legend}{Description.}

\item{inherit.aes}{Description.}

\item{label}{Description.}

\item{type}{Description.}

\item{xmin}{Description.}

\item{xmax}{Description.}

\item{y.position}{Description.}

\item{label.size}{Description.}

\item{tip.length}{Description.}

\item{bracket.size}{Description.}

\item{bracket.colour}{Description.}

\item{bracket.shorten}{Description.}

\item{bracket.nudge.y}{Description.}

\item{step.increase}{Description.}

\item{step.group.by}{Description.}

\item{coord.flip}{Description.}

\item{...}{Description.}
}
\value{
Description.
}
\description{
Description.
}
\examples{
#

}
