% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_ticks.R, R/geom_bracket.R
\docType{data}
\name{GeomTicks}
\alias{GeomTicks}
\alias{GeomBracket}
\title{ggproto classes for ggprism}
\description{
ggproto classes for ggprism
}
\keyword{datasets}
\keyword{internal}
