% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggprism-package.R
\docType{package}
\name{ggprism-package}
\alias{ggprism}
\alias{ggprism-package}
\title{ggprism: GraphPad Prism Inspired Themes and Palettes for ggplot2}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Charlotte Dawson \email{csdaw@outlook.com} (\href{https://orcid.org/0000-0002-7151-5971}{ORCID})

}
\keyword{internal}
