% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggprism_data.R
\docType{data}
\name{ggprism_data}
\alias{ggprism_data}
\title{Palettes and theme data for ggprism}
\format{
An object of class \code{list} of length 4.
}
\usage{
ggprism_data
}
\description{
This list object contains the strings and values used in
ggprism themes and palettes.
}
\keyword{datasets}
