% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_prism_minor.R
\name{guide_train.prism_minor}
\alias{guide_train.prism_minor}
\alias{guide_gengrob.prism_minor}
\title{Helper methods for guides}
\usage{
\method{guide_train}{prism_minor}(guide, scale, aesthetic = NULL)

\method{guide_gengrob}{prism_minor}(guide, theme)
}
\description{
Helper methods for guides
}
\keyword{internal}
