% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_prism_minor.R
\name{guide_gengrob.prism_minor}
\alias{guide_gengrob.prism_minor}
\title{Title}
\usage{
guide_gengrob.prism_minor(guide, theme)
}
\arguments{
\item{guide}{Description.}

\item{theme}{Description.}
}
\value{

}
\description{
Title
}
