% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_prism_bracket.R
\name{guide_prism_bracket}
\alias{guide_prism_bracket}
\title{Axis guide}
\usage{
guide_prism_bracket(
  title = waiver(),
  check.overlap = FALSE,
  angle = NULL,
  n.dodge = 1,
  order = 0,
  position = waiver(),
  bracket_width = 0.1,
  outside = TRUE
)
}
\arguments{
\item{title}{Description.}

\item{check.overlap}{silently remove overlapping labels,
(recursively) prioritizing the first, last, and middle labels.}

\item{angle}{Compared to setting the angle in [theme()] / [element_text()],
this also uses some heuristics to automatically pick the `hjust` and `vjust` that
you probably want.}

\item{n.dodge}{The number of rows (for vertical axes) or columns (for
horizontal axes) that should be used to render the labels. This is
useful for displaying labels that would otherwise overlap.}

\item{order}{Used to determine the order of the guides (left-to-right,
top-to-bottom), if more than one  guide must be drawn at the same location.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}

\item{bracket_width}{Description.}

\item{outside}{Description.}
}
\description{
Axis guides are the visual representation of position scales like those
created with [scale_(x|y)_continuous()][scale_x_continuous()] and
[scale_(x|y)_discrete()][scale_x_discrete()].
}
\examples{
#

}
