% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_prism_offset_minor.R
\name{guide_prism_offset_minor}
\alias{guide_prism_offset_minor}
\title{Offset axis guide with minor ticks}
\usage{
guide_prism_offset_minor(
  title = waiver(),
  check.overlap = FALSE,
  angle = NULL,
  n.dodge = 1,
  order = 0,
  position = waiver()
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{check.overlap}{silently remove overlapping labels,
(recursively) prioritizing the first, last, and middle labels.}

\item{angle}{Compared to setting the angle in \code{\link[ggplot2:theme]{theme()}} / \code{\link[ggplot2:element]{element_text()}},
this also uses some heuristics to automatically pick the \code{hjust} and \code{vjust} that
you probably want.}

\item{n.dodge}{The number of rows (for vertical axes) or columns (for
horizontal axes) that should be used to render the labels. This is
useful for displaying labels that would otherwise overlap.}

\item{order}{Used to determine the order of the guides (left-to-right,
top-to-bottom), if more than one  guide must be drawn at the same location.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}
}
\description{
This guide draws the axis only as wide as the outermost tick marks,
similar to offset axes from Prism. It also adds minor ticks.
}
\details{
Control the length of the axis by adjusting the \code{breaks} argument in
\verb{scale_(x|y)_continuous()} or \verb{scale_(x|y)_discrete()}. Similarly,
the number of minor ticks can be changed using the \code{minor_breaks}
argument.

Control the length of minor ticks by setting \code{prism.ticks.length} to
a \code{\link[grid]{unit}} object using \code{\link[ggplot2]{theme}},
for example: \code{prism.ticks.length = unit(2, "pt")}. The major tick
lengths are adjusted using the standard \code{axis.ticks.length}.
}
\examples{
# example
}
