% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_prism_minor.R
\name{guide_train.prism_minor}
\alias{guide_train.prism_minor}
\title{Title}
\usage{
\method{guide_train}{prism_minor}(guide, scale, aesthetic = NULL)
}
\arguments{
\item{guide}{Description.}

\item{scale}{Description.}

\item{aesthetic}{Description.}
}
\value{

}
\description{
Title
}
\keyword{internal}
