% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_theme.R
\name{preview_theme}
\alias{preview_theme}
\title{Preview Prism themes}
\usage{
preview_theme(palette)
}
\arguments{
\item{palette}{\code{string}. Palette name.}
}
\description{
Quickly generate a preview of a ggprism theme.
See \code{names(ggprism_data$themes)} for valid palette names.
}
\examples{
library(ggplot2)

## see names of available themes
names(ggprism_data$themes)

## preview a theme
preview_theme("floral")
}
