% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_prism.R
\name{prism_colour_pal}
\alias{prism_colour_pal}
\alias{prism_color_pal}
\title{Prism colour palettes (discrete)}
\usage{
prism_colour_pal(palette = "colors")

prism_color_pal(palette = "colors")
}
\arguments{
\item{palette}{string. Palette name, see names(ggprism_data$colour_palettes)
for valid palette names.}
}
\description{
Description.
}
\examples{
# example goes here
}
\seealso{
Other colour prism: 
\code{\link{scale_colour_prism}()}
}
\concept{colour prism}
