% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_shape_prism.R
\name{prism_shape_pal}
\alias{prism_shape_pal}
\title{Prism shape palettes (discrete)}
\usage{
prism_shape_pal(palette = c("default", "filled", "complete"))
}
\arguments{
\item{palette}{string. Palette name, one of: default, filled,
or complete.}
}
\description{
Shape palettes used by
\href{https://www.graphpad.com}{GraphPad Prism}
}
\details{
The complete palette is not exact but approximates
the shape palette from Prism using only the default symbols
available in R. No unicode characters are used.

The default palette supports up to 9 values. It does not use
any symbols with a fill.

The filled palette supports up to 10 values. The first 5 symbols
have a fill.

The complete palette supports up to 14 values. Symbols 5 to 9
have a fill.
}
\examples{
# need to write an example
}
\concept{shape}
