% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_prism.R
\name{theme_prism}
\alias{theme_prism}
\title{Prism themes}
\usage{
theme_prism(
  palette = "black_and_white",
  base_size = 12,
  base_family = "sans",
  base_fontface = "bold",
  base_line_size = base_size/20,
  base_rect_size = base_size/20,
  axis_text_angle = 0,
  border = FALSE
)
}
\arguments{
\item{palette}{`string`. Palette name, use
`names(ggprism_data$themes)` to show all valid palette names.}

\item{base_size}{`numeric`. Base font size, given in `"pt"`.}

\item{base_family}{`string`. Base font family, default is `"sans"`.}

\item{base_fontface}{`string`. Base font face, default is `"bold"`.}

\item{base_line_size}{`numeric`. Base linewidth for line elements}

\item{base_rect_size}{`numeric`. Base linewidth for rect elements}

\item{axis_text_angle}{`integer`. Angle of axis text in degrees.
One of: `0, 45, 90, 270`.}

\item{border}{`logical`. Should a border be drawn around the plot?
Clipping will occur unless e.g. `coord_cartesian(clip = "off")` is used.}
}
\description{
A collection of ggplot2 themes that use palettes which mirror the
colour schemes available in GraphPad Prism.
}
\examples{
# example goes here
}
