% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pvalue.R
\name{add_pvalue}
\alias{add_pvalue}
\title{Add p-values to a ggplot}
\usage{
add_pvalue(
  data,
  label = NULL,
  xmin = "group1",
  xmax = "group2",
  x = NULL,
  y.position = "y.position",
  label.size = 3.2,
  colour = NULL,
  color = NULL,
  tip.length = 0.03,
  bracket.size = 0.6,
  bracket.colour = NULL,
  bracket.shorten = 0,
  bracket.nudge.y = 0,
  step.increase = 0,
  step.group.by = NULL,
  remove.bracket = FALSE,
  coord.flip = FALSE,
  position = "identity",
  ...
)
}
\arguments{
\item{data}{A `data.frame` with the statistics to plot. Default format
has the following columns: `group1 | group2 | p.adj | y.position | etc`.
`group1` and `group2` are the two groups that were compared.
`p.adj` is the adjusted p-value. `y.position` is the y coordinate
that specifies where on the plot the p-value should go. The column names
can differ from the default as long as their are specified when calling
the function.}

\item{label}{`string`. Name of column in `data` that
contains the text to plot (e.g. `label = "p.adj"`).
Can also be an expression that can be formatted by
\code{\link[glue]{glue}} (e.g. `label = "p = {p.adj}"`).}

\item{xmin}{`string`. Name of column in `data`  that
contains the position of the left side of the brackets.
Default is `"group1"`.}

\item{xmax}{Optional. `string`. Name of column in `data` that
contains the position of the right side of the brackets.
Default is `"group2"`. If `NULL`, the p-values are plotted as
text only.}

\item{x}{`string` or `numeric`. x coordinate of the p-value text.
Only use when plotting p-value as text only (no brackets).}

\item{y.position}{`string`. Name of column in
\code{data} containing the y coordinates (`numeric`) of each p-value to
be plotted. Can also be a single number to plot all p-values at the same
height or a `numeric` vector that will override `data`.}

\item{label.size}{`numeric`. Size of text.}

\item{colour, color}{`string`. Colour of text.}

\item{tip.length}{`numeric` vector. Length of bracket tips.
Use `0` to remove tips.}

\item{bracket.size}{`numeric`. Linewidth of bracket.}

\item{bracket.colour}{`string`. Colour of bracket. Default is `NULL`
which causes brackets to inherit the colour of the text.}

\item{bracket.shorten}{`numeric`. Shortens the brackets slightly to
allow them to be plotted side-by-side at the same y position.}

\item{bracket.nudge.y}{`numeric`. Changes the y position of
p-values. Useful for slightly adjusting p-values if the text is cut off.}

\item{step.increase}{`numeric`. Changes the space between brackets.}

\item{step.group.by}{`string`. Variable to group brackets by.}

\item{remove.bracket}{`logical`. If `TRUE` all brackets are
removed and p-value is shown as text only.}

\item{coord.flip}{`logical`. If `TRUE` p-values are rotated by
90 degrees. Should be used with \code{\link[ggplot2]{coord_flip}}}

\item{position}{`string` or call to position function such as
\code{\link[ggplot2]{position_dodge}}. Typically used for adjusting x
position of p-values to be in line with dodged data.}

\item{...}{Additional aesthetics or arguments passed to
\code{\link[ggplot2]{layer}}. See below for allowed values.}
}
\description{
Add p-values with or without brackets to a ggplot.

See **here** or the examples section below for examples of how to use.

`add_pvalue` is a refactored version of
`stat_pvalue_manual` from
[kassambara/ggpubr](https://github.com/kassambara/ggpubr), altered to
have less dependencies, and more flexibility with input format and
aesthetics. Any examples using `stat_pvalue_manual` found on
[Datanovia](https://www.datanovia.com/en/?s=p-value&search-type=default)
will also work with `add_pvalue`.
}
\section{Allowed ... values}{

`add_pvalue` understands the following additional aesthetics or arguments:

\describe{
  \item{`fontface`}{`string`. Fontface of text (e.g. `"bold"`).}
  \item{`fontfamily`}{`string`. Fontfamily of text (e.g. `"Arial"`).}
  \item{`hjust`}{`numeric`. Horizontal justification of text.}
  \item{`vjust`}{`numeric`. Vertical justification of text.}
  \item{`alpha`}{`numeric`. Transparency of text and/or brackets.}
  \item{`linetype`}{`string` or `numeric`. Linetype of brackets
   (e.g. `"dashed"`).}
  \item{`lineend`}{`string`. Lineend of brackets (e.g. `"butt"`).}
  \item{`na.rm`}{`logical`. If `FALSE` (default), removes
  missing values with a warning. If `TRUE` silently removes missing
  values.}
  \item{`show.legend`}{`logical`. Should this layer be included in
  the legends? If `NA` (default), include if any aesthetics are mapped.
  If `FALSE`, never include or if `TRUE`, always include. It can
  also be a named `logical` vector to finely select the aesthetics to
  display.}
  \item{`inherit.aes`}{`logical`. If `FALSE`, overrides the
  default aesthetics, rather than combining with them.}
}
}

\examples{
#

}
