% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pvalue.R, R/geom_bracket.R
\name{add_pvalue}
\alias{add_pvalue}
\alias{geom_bracket}
\title{Add p-values to a ggplot}
\usage{
add_pvalue(
  data,
  label = NULL,
  xmin = "group1",
  xmax = "group2",
  x = NULL,
  y.position = "y.position",
  label.size = 3.2,
  colour = NULL,
  color = NULL,
  tip.length = 0.03,
  bracket.size = 0.6,
  bracket.colour = NULL,
  bracket.shorten = 0,
  bracket.nudge.y = 0,
  step.increase = 0,
  step.group.by = NULL,
  remove.bracket = FALSE,
  coord.flip = FALSE,
  position = "identity",
  ...
)

geom_bracket(
  mapping = NULL,
  data = NULL,
  stat = "bracket",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  label = NULL,
  type = c("text", "expression"),
  xmin = NULL,
  xmax = NULL,
  y.position = NULL,
  label.size = 3.2,
  tip.length = 0.03,
  bracket.size = 0.6,
  bracket.colour = NULL,
  bracket.shorten = 0,
  bracket.nudge.y = 0,
  step.increase = 0,
  step.group.by = NULL,
  coord.flip = FALSE,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} with the statistics to plot. Default format
has the following columns: \code{group1 | group2 | p.adj | y.position | etc}.
\code{group1} and \code{group2} are the two groups that were compared.
\code{p.adj} is the adjusted p-value. \code{y.position} is the y coordinate
that specifies where on the plot the p-value should go. The column names
can differ from the default as long as their are specified when calling
the function.}

\item{label}{\code{string}. Name of column in \code{data} that
contains the text to plot (e.g. \code{label = "p.adj"}).
Can also be an expression that can be formatted by
\code{\link[glue]{glue}} (e.g. \code{label = "p = {p.adj}"}).}

\item{xmin}{\code{string}. Name of column in \code{data}  that
contains the position of the left side of the brackets.
Default is \code{"group1"}.}

\item{xmax}{Optional. \code{string}. Name of column in \code{data} that
contains the position of the right side of the brackets.
Default is \code{"group2"}. If \code{NULL}, the p-values are plotted as
text only.}

\item{x}{\code{string} or \code{numeric}. x coordinate of the p-value text.
Only use when plotting p-value as text only (no brackets).}

\item{y.position}{\code{string}. Name of column in
\code{data} containing the y coordinates (\code{numeric}) of each p-value to
be plotted. Can also be a single number to plot all p-values at the same
height or a \code{numeric} vector that will override \code{data}.}

\item{label.size}{\code{numeric}. Size of text.}

\item{colour, color}{\code{string}. Colour of text.}

\item{tip.length}{\code{numeric} vector. Length of bracket tips.
Use \code{0} to remove tips.}

\item{bracket.size}{\code{numeric}. Linewidth of bracket.}

\item{bracket.colour}{\code{string}. Colour of bracket. Default is \code{NULL}
which causes brackets to inherit the colour of the text.}

\item{bracket.shorten}{\code{numeric}. Shortens the brackets slightly to
allow them to be plotted side-by-side at the same y position.}

\item{bracket.nudge.y}{\code{numeric}. Changes the y position of
p-values. Useful for slightly adjusting p-values if the text is cut off.}

\item{step.increase}{\code{numeric}. Changes the space between brackets.}

\item{step.group.by}{\code{string}. Variable to group brackets by.}

\item{remove.bracket}{\code{logical}. If \code{TRUE} all brackets are
removed and p-value is shown as text only.}

\item{coord.flip}{\code{logical}. If \code{TRUE} p-values are rotated by
90 degrees. Should be used with \code{\link[ggplot2]{coord_flip}}}

\item{position}{\code{string} or call to position function such as
\code{\link[ggplot2]{position_dodge}}. Typically used for adjusting x
position of p-values to be in line with dodged data.}

\item{...}{Additional aesthetics or arguments passed to
\code{\link[ggplot2]{layer}}. See below for allowed values.}
}
\description{
Add p-values with or without brackets to a ggplot.

See \strong{here} or the examples section below for examples of how to use.

\code{add_pvalue} is a refactored version of
\code{stat_pvalue_manual} from
\href{https://github.com/kassambara/ggpubr}{kassambara/ggpubr}, altered to
have less dependencies, and more flexibility with input format and
aesthetics. Any examples using \code{stat_pvalue_manual} found on
\href{https://www.datanovia.com/en/?s=p-value&search-type=default}{Datanovia}
will also work with \code{add_pvalue}.
}
\section{Allowed ... values}{

\code{add_pvalue} understands the following additional aesthetics or arguments:

\describe{
  \item{\code{fontface}}{\code{string}. Fontface of text (e.g. \code{"bold"}).}
  \item{\code{fontfamily}}{\code{string}. Fontfamily of text (e.g. \code{"Arial"}).}
  \item{\code{hjust}}{\code{numeric}. Horizontal justification of text.}
  \item{\code{vjust}}{\code{numeric}. Vertical justification of text.}
  \item{\code{alpha}}{\code{numeric}. Transparency of text and/or brackets.}
  \item{\code{linetype}}{\code{string} or \code{numeric}. Linetype of brackets
   (e.g. \code{"dashed"}).}
  \item{\code{lineend}}{\code{string}. Lineend of brackets (e.g. \code{"butt"}).}
  \item{\code{na.rm}}{\code{logical}. If \code{FALSE} (default), removes
  missing values with a warning. If \code{TRUE} silently removes missing
  values.}
  \item{\code{show.legend}}{\code{logical}. Should this layer be included in
  the legends? If \code{NA} (default), include if any aesthetics are mapped.
  If \code{FALSE}, never include or if \code{TRUE}, always include. It can
  also be a named \code{logical} vector to finely select the aesthetics to
  display.}
  \item{\code{inherit.aes}}{\code{logical}. If \code{FALSE}, overrides the
  default aesthetics, rather than combining with them.}
}
}

\examples{
#

}
