% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_ticks.R
\name{annotation_ticks}
\alias{annotation_ticks}
\title{Title}
\usage{
annotation_ticks(
  sides = "b",
  type = "both",
  scale = "identity",
  scaled = TRUE,
  outside = FALSE,
  ticklength = unit(5.5, "pt"),
  minorlength = unit(2.4, "pt"),
  colour = "black",
  base_size = 12,
  linetype = 1,
  lineend = "butt",
  alpha = 1,
  color = NULL,
  ticks_per_base = NULL,
  data = data.frame(x = NA),
  ...
)
}
\arguments{
\item{sides}{Description.}

\item{scale}{Description.}

\item{scaled}{Description.}

\item{outside}{Description.}

\item{ticklength}{Description.}

\item{minorlength}{Description.}

\item{colour}{Description.}

\item{base_size}{Description.}

\item{linetype}{Description.}

\item{alpha}{Description.}

\item{color}{Description.}

\item{ticks_per_base}{Description.}

\item{data}{Description.}

\item{...}{Description.}
}
\value{
Description.
}
\description{
Code modified from teunbrand's answer to https://stackoverflow.com/questions/58485334
which is itself a refactored version of hrbrmstr's annotation_ticks function.
https://github.com/hrbrmstr/ggalt/blob/master/R/annotation_ticks.r
}
\examples{
# example goes here
}
