% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_ticks.R, R/geom_bracket.R
\docType{data}
\name{GeomTicks}
\alias{GeomTicks}
\alias{StatBracket}
\alias{GeomBracket}
\title{ggproto classes for ggprism}
\format{
An object of class \code{StatBracket} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomBracket} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
StatBracket

GeomBracket
}
\description{
ggproto classes for ggprism
}
\keyword{datasets}
\keyword{internal}
