% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggprism-package.R
\docType{package}
\name{ggprism-package}
\alias{ggprism}
\alias{ggprism-package}
\title{ggprism: A ggplot2 Extension Inspired by GraphPad Prism}
\description{
Provides various themes, palettes, and other functions
    that are used to customise ggplots to look like they were made in GraphPad 
    Prism. The 'Prism-look' is achieved with theme_prism() and 
    scale_fill|colour_prism(), axes can be changed with custom guides like 
    guide_prism_minor(), and significance indicators added with add_pvalue().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://csdaw.github.io/ggprism/}
  \item \url{https://github.com/csdaw/ggprism}
  \item Report bugs at \url{https://github.com/csdaw/ggprism/issues}
}

}
\author{
\strong{Maintainer}: Charlotte Dawson \email{csdaw@outlook.com} (\href{https://orcid.org/0000-0002-7151-5971}{ORCID})

}
\keyword{internal}
