% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggprism_data.R
\docType{data}
\name{ggprism_data}
\alias{ggprism_data}
\title{Title}
\format{
An object of class \code{list} of length 3.
}
\usage{
ggprism_data
}
\value{
Description.
}
\description{
Description.
}
\keyword{datasets}
