% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_prism_bracket.R
\name{guide_prism_bracket}
\alias{guide_prism_bracket}
\title{Axis guide with brackets}
\usage{
guide_prism_bracket(
  title = waiver(),
  check.overlap = FALSE,
  angle = NULL,
  n.dodge = 1,
  order = 0,
  position = waiver(),
  bracket_width = NULL,
  outside = TRUE
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{check.overlap}{silently remove overlapping labels,
(recursively) prioritizing the first, last, and middle labels.}

\item{angle}{Compared to setting the angle in \code{\link[ggplot2:theme]{theme()}} / \code{\link[ggplot2:element]{element_text()}},
this also uses some heuristics to automatically pick the \code{hjust} and \code{vjust} that
you probably want.}

\item{n.dodge}{The number of rows (for vertical axes) or columns (for
horizontal axes) that should be used to render the labels. This is
useful for displaying labels that would otherwise overlap.}

\item{order}{Used to determine the order of the guides (left-to-right,
top-to-bottom), if more than one  guide must be drawn at the same location.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}

\item{bracket_width}{\code{numeric}. Controls the width of the bracket. Try
values between 0 and 1.}

\item{outside}{\code{logical}. Default is \code{TRUE} and brackets point
outwards. If \code{FALSE} the bracket crossbar is moved so the ticks appear
to point inwards towards the plotting area.}
}
\description{
This guide turns the axis into brackets drawn around each axis label.
}
\details{
The number of brackets can be adjusted using the \code{breaks}
argument in \code{scale_(x|y)_continuous()} or \code{scale_(x|y)_discrete()}.
}
\examples{
# example
}
