% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_prism_minor.R
\name{guide_prism_minor}
\alias{guide_prism_minor}
\title{Axis guide}
\usage{
guide_prism_minor(
  title = waiver(),
  check.overlap = FALSE,
  angle = NULL,
  n.dodge = 1,
  order = 0,
  position = waiver()
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If `NULL`, the title is not shown. By default
([waiver()]), the name of the scale object or the name
specified in [labs()] is used for the title.}

\item{check.overlap}{silently remove overlapping labels,
(recursively) prioritizing the first, last, and middle labels.}

\item{angle}{Compared to setting the angle in [theme()] / [element_text()],
this also uses some heuristics to automatically pick the `hjust` and `vjust` that
you probably want.}

\item{n.dodge}{The number of rows (for vertical axes) or columns (for
horizontal axes) that should be used to render the labels. This is
useful for displaying labels that would otherwise overlap.}

\item{order}{Used to determine the order of the guides (left-to-right,
top-to-bottom), if more than one  guide must be drawn at the same location.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}
}
\description{
Axis guides are the visual representation of position scales like those
created with [scale_(x|y)_continuous()][scale_x_continuous()] and
[scale_(x|y)_discrete()][scale_x_discrete()].
}
\examples{
# example


}
