% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_prism.R
\name{prism_fill_pal}
\alias{prism_fill_pal}
\title{Prism fill palettes (discrete)}
\usage{
prism_fill_pal(palette = "colors")
}
\arguments{
\item{palette}{string. Palette name, see names(ggprism_data$fill_palettes)
for valid palette names.}
}
\description{
Description.
}
\examples{
# need to write an example
}
\concept{fill}
