% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_prism.R
\name{theme_prism}
\alias{theme_prism}
\title{Prism themes}
\usage{
theme_prism(
  palette = "black_and_white",
  base_size = 12,
  base_family = "sans",
  base_fontface = "bold",
  base_line_size = base_size/20,
  base_rect_size = base_size/20,
  axis_text_angle = 0,
  border = FALSE
)
}
\arguments{
\item{palette}{\code{string}. Palette name, use
\code{names(ggprism_data$themes)} to show all valid palette names.}

\item{base_size}{\code{numeric}. Base font size, given in \code{"pt"}.}

\item{base_family}{\code{string}. Base font family, default is \code{"sans"}.}

\item{base_fontface}{\code{string}. Base font face, default is \code{"bold"}.}

\item{base_line_size}{\code{numeric}. Base linewidth for line elements}

\item{base_rect_size}{\code{numeric}. Base linewidth for rect elements}

\item{axis_text_angle}{\code{integer}. Angle of axis text in degrees.
One of: \verb{0, 45, 90, 270}.}

\item{border}{\code{logical}. Should a border be drawn around the plot?
Clipping will occur unless e.g. \code{coord_cartesian(clip = "off")} is used.}
}
\description{
A collection of ggplot2 themes that use palettes which mirror the
colour schemes available in GraphPad Prism.
}
\examples{
# example goes here
}
