% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_prism.R
\name{theme_prism}
\alias{theme_prism}
\title{Title}
\usage{
theme_prism(
  palette = "black_and_white",
  base_size = 12,
  base_family = "",
  base_fontface = "bold",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  axis_text_angle = 0,
  frame = FALSE
)
}
\arguments{
\item{palette}{Description.}

\item{base_size}{Description.}

\item{base_family}{Description.}

\item{base_fontface}{Description.}

\item{base_line_size}{Description.}

\item{base_rect_size}{Description.}

\item{axis_text_angle}{Description.}

\item{frame}{Description.}
}
\value{

}
\description{
Description.
}
\examples{
# example goes here
}
